Fixed bug with static state in HighlightingService
HighlightingEditorView was crashing after opening CodeRunnerView because HighlightingService used a static property (`CurrentXshdXml`) to store the current XSHD definition. CodeRunnerView would set the definition to `asm.xshd`, which would cause HighlightingEditorView to try to edit the wrong definition, causing the application to crash.

This issue was resolved by removing the `CurrentXshdXml` static property and introducing a new `GetXshdXml` method that retrieves the correct XSHD content on demand for a specific language and theme.